Vim: removed some airline settings:
Setting to show buffers filenames at the top of the screen was, due
to a bug, causing extra text to get inserted at into some buffers.
Also, I ended up realizing it is that useful of a feature, possibly
even distracting, and I just liked it because it looks cool. I say
distracting because I use bufexplorer.