Better handling of tags:
Using vim-tags plugin. I tried vim-easytags, but it was anyting but
easy and didn't really work. vim-tags is not perfect, but it is the
best tag solution I have found so far. It looks for indications of
project root directory, such as a .git directory, and creates tag
files there. First, you run :TagsGenerate! to create a tags file
for a project. Then :TagsGenerate is used automatically to keep
the tags file updated every time you save. I am having problems
with a bug where :TagsGenerate does not produce all the needed,
but :TagsGenerate! does, so for now, I am running that manually
when needed. I created an issue for this. If it does not get
attention relatively quickly, I will try to create a customization
where :TagsGenerate! gets run automatically.