Added an awesome Fp command to search through .php files:
This searches through .php files without using regular expressions.
It is the Vim equivalent of calling my shell function, fp. You just
enter the command :Fp followed by what you want to search for. You
don't need to type quotes around what you are searching for. This
will create a scratch buffer with the results. Once inside the
scratch buffer the following rules apply. Press q to discard the
search results and get back to what you were doing. Pressing Enter
will fill go to that search result. It will open that file in the
window above the search results and jump to the correct line in the
file automatically.