Vim: Slight improvement to php search:
Now insided the search results, the search term will be jumped to,
and other matches in that window will be highlighted. Pressing n
will jump to the next match in the search results window/buffer.
One exception is if this is the use has not done any searching yet,
the results in the search window will not have any highlighting.
However, the cursor will still have moved to the first instance of
the search query.