Vim: Started making tags more accurate:
Remived the tags plugin I was using. Removed the settings for it.
Started making some functions that will be ultimately help looking
up a tag to be more accuate. Wrote a couple sh scripts that will be
used for the same purpose. None of this is working yet, but the
exisint code is useful and going down a good path.