Vim: Improved tag experience in php files:
Tags are generated automatically upon saving a php file. This tags
file will only contain php tags. It is named phptags. The project
directory is calculated based on where the .git directory is, and
the tags file is saved in your project root directory. It is not
saved in the .git directory. When a viewing php file, Vim only
looks inside the phptags file for the current project. This
goes a long way to preventing false positives with matching tags.
The same thing can be setup for other languages. Soon I would like
to have JavaScript files working this way too. In that case, the
same project could contain both a phptags file and a jstags file at
the project root directory, and the correct file would be used for
both looking up and saving tags depending on what type of file you
are editing. This does not work well for files that have a mix of
multiple languages inside them. I am hoping this will be a good
thing and encourage me to keep my code separated.