Vim: Removed YouCompleteMe plugin
It has been causing lots of problems lately. It is hard to compile
for a server a work from sometimes. But I am still using Neovim on
that server, so I always get this error message from YouCompleteMe
every time I start Neovim. When I am using it on my local machine,
lately it has been causing problems. When it gets activated,
Neovim just hangs. I have to press Control+C to back out of it.
Also, it often moves my buffer view so that the line the cursor is
on is at the top of the window, which is very jarring and
disruptive. When I am trying to record a macro, it can get
activated and mess things up. I will probably install it again and
play with the settings at some point, but for now I am removing it.