Vim: Created a plugin/library called elhiv:
elhiv stands for extreme language hacking, which is my programming
style. But it is just a VimL library. My vim-project-tags plugin
depends on it. I will start moving general purpose functions into
it. It uses vim-UT for automated tests, so I added that plugin and
its dependencies to my config.