Changed path back to older ruby version:
speckle now works correctly both on my computer and on dev. On dev
I had to play around with different versions of ruby and rubygems,
but I finally got it working. The version of ruby referenced here
is there the gem binaries are installed, although I'm not sure why
as dev actually is using a higher version of ruby than what is
indicated by this directory name. Oh well it all seems to be
working.