Vim: config setting to not be a developer of my plugins:
This allows myself or others to use these dot files on a machine
where development on my plugins does not happen. This makes plug
use the master branch of these plugins. It also prevents you from
needing to have ssh keys setup that are authorized to work with my
plugins.