Neovim: Added a MochaFilter command:
It runs tests matching what is under the cursor. If the cursor is
on a describe line, it runs all the tests with that describe text,
so all the tests for that method or propertly. If the cursor is at
an it line, it just runs tests matching that it text, so probably
just that one desribe test for that one method/property.