Vim: Changed the mappings:
The mappings for moving parameters around are very rarely used
because that functionality mostly does not work as desired. So I
changed those mappings to have 2 leaders in front of them instead
of 1 in order to free up the 1 leader mappings to use for something
that I will use more often.

<leader>< is now mapped to the new Less command
<leader>> is now mapped to the new More command
<leader>= is now mapped to the new Same command

Using those mappings is much more convenient than typing those
commands.