Vim: For tab indented files, allowing C style multi-line comments to not be considered mixed-indent:
Evenatually when I am using better tools to check this, I can
completely disable that from within airline, at least for certain
file types. For example I will eventually use phpcs for this and
integrat that within neomake. But for now ariline is the only
thing I am using. This setting is not ideal at all, but I would
rather have it tell me everything is ok when it is not than tell
me there are problems when there are not, so this is an
improvement.