Vim config: improved PHP paste mappings:
The paste template now allows you to specify a default class to
inherit from. A good place to set that is in the .afterinit.vim
file.

The paste method mapping is now smarter. As long as you run it from
a non-blank line with the intention of pasting after that line, it
will create the correct blank lines before and after the method if
and where appropriate.