Trying to leave .bashrc alone:
This will take multiple commits, but the idea is you can use
whatever .bashrc your system comes with and just add a line to
source the .dreamybashrc file at the end. This may or may not work
out well, but I want to try out the default .bashrc in conjunction
with this one on a new system I am working from.