gitignoring the .bashrc file:
This is the second part of a 2 part commit. The idea is to allow
the user to keep their original .bashrc file mostly untouched. You
can either manually replace the entire contents of your .bashrc
with a line sourcing the .dreamy bashrc or place that line at the
end of your .bashrc file.