Vim: Added a global config value for a log file:
There was already a logging system in place used by my plugins and
Vim code that allowed you to specify a log file. However, it was
previously not possible to set this from inside the .beforeinit.vim
file because it had to be an object for a class that was not yet
defined. Now there is an additional config value called
g:dreamy_log which can be set in the .beforeinit.vim file to a
string path. That value is later used to set the already used log
config value to an appropriate L_file object. If you don't set this
new config value to anything, logging will be off by default. This
allows me to turn on logging on different computers I use and
specify different log paths on them without turning on logging by
default and wihout having a default log path.