leader V works for php as well:
leader V (as in capital V) is similar to leader v except that it is
supposed to run tests across the entire project, instead of just for the
file you are viwing. Now if you run this command from a php, it will run
the 'composer test' command. You will have to configure your
composer.json file to have a custom 'test' command that runs your tests
in whatever way is appropriate for that project. This is a great way of
testing because composer temporarily adds the local vendor/bin
directory to the highest priority spot in your $PATH.