normal mode mapping <leader>@ replays the last played macro 3 times:
Once I am wanting to replay the last played macro, it is very common for
me to want to play it many times. But normally I don't know exactly how
many times I want to play it, I just keep playing it until the text is
right. <leader>@ is a natural synonym to @@, which replays the last
played macro. I played around with different numbers for how many times
this should replay the last macro. Three seems to be the best number as
it is much faster than just pressing @@ over and over, and it minizes
the number of times @@ will have to be pressed at the very end of the
sequence when 3 is not the exact number of replays needed to finish off
the editing.