Added p as a bash alias to composer:
I am using composer a lot lately, even for running commands that don't
have anything to do with composer. Well, sort of. Composer has this
feature where you can write commands specific to a project into your
composer.json file, and then run them via "composer my-command". They
utilitze the local vendor/bin directory automatically.

Anyway, my point is, I have good reason to be typing "composer" a whole
lot lately, more so than what might seem obviously practical. And I
misspell it a lot, and it is kind of long for a command name, so as much
as it might seem silly, p as an alias to composer is actually really
nice.

I could have used c, but that is already an I use for cd. I could have
done something like comp or poser, but p is much shorter. I believe in
readability over brievity for variable names and code, but not for
aliases for frequently typed commands.