nvim improved lookup definition:
When using Control-] to lookup the definition of something, I have this
configued to use the language server lookup feature for php files. Now
it also the current location to the jump list before looking it up so
you can easilly jump back to where you were before the lookup.