alias improvements:
Added the 'c.' alias back again because it turns out that 'cd' is not
the same as 'cd ..', it is actualy the same as 'cd ~'. So 'c.' is back
and once again is an alias for 'cd ..'.

On a related note, I removed the 'c~' alias since that is the one that
is actually the same as 'c', since that is an alias for 'cd', which by itself
will do a 'cd ~'

In review, use 'c' to go home, and 'c.' to go up a directory.