nvim: removed command to navigate to vim-project-tags:
I am even more convinced than ever before that tags are unlikely to be a
valuable avenue to pursue or think about. I recently became really
frustrated with code lookup options for Neovim, and I thought maybe tags
would work better for me now since our PHP code is more organized and
uses proper namespaces and file locations and lots of type information.
I tried out my plugina again, and was reminded of why I abondoned tags
to begin with. The problem is not the generation of tags, but the lookup
of tags. The same parser that is used to generate the tags, needs to
become activated at the time a tag is looked up to decide how to lookup
the tag. But that is not how it works. Even if that is how it worked, it
seems like it still wouldn't make sense because then tags are no longer
needed. Why activate a parser cabable of generating tags and have it do
all the work necessary to generate tags, but then instead only decide
how to search through the tags for a specific tag? Instead it would make
more sense for the parse to just provide whatever information you are
inquiring about directly.