added .ideavimrc to tracked dot files and got IDEAVim using some of my shared Vim settings:
Now the init.core.vim has more settings in it that have been moved
out of init.vim. Both Neovim and IDEAVim use and are compatible
with the code in .ideavimrc. In .ideavimrc, global variables are
not set using g:variable_name because for some reason it was only
working without the leading g:. I'm not sure if that is always the
case or if it only applied to mapleader...