A bunch of improvements:
No longer using deoplete. It was causing problems AGAIN so I tried
disabling it, and for some reason I am now happier with with manually
invoking the builtin nvim completion when desired.

Removed all rainbow parenthesis stuff because NONE of the plugins out
there work with PHP code.

Changed the default PHP paste template to not put
declare(strict_types=1); at the top of the file because for most
projects it makes more sense to let Psalm strictly enforce the types
instead of PHP. This way you are favoring lint time errors over runtime
errors. The way Psalm deals with types if much more configurable than
PHP's option of just turning strict types on or off, so it is much more
practical for PHP code bases, which will have technical debt. This is
always true because just the fact that it is PHP code is itself
technical debt, but part of cleaning up that code can involve preparing
the code to be transpiled or rewritten in another language (like
Kotlin), using Psalm to help you do this is a better strategy than using
PHP's strict types option for the reasons I am explaining (yes this is
circular logic, get over it by thinking of it as recursion).

Added a new command called Psalmpress which calls a new method called
DreamyPsalmPress. The command/method accept a single parameter, which is
the name of the Psalm issue type you wish to suppress. This is very
helpful because Psalm issues often need to be temporarilly suppressed
until you can clean up the code later. This is partly because new Psalm
releases keep including bug fixes or new issue types that may expose
type issues you were not ready to deal with.

Changed the default PHP test class template to suppress Psalm's
UnusedClass issue, since it can appear that test classes are unused even
though they are not.