Freed up unused vim mapping to edit one off vim functions file:
I eneded up never using this file, and checked and found out it does not
exist on any of my systems. This is a nice mapping to have freed up and
available for use, and I am using it on one of my systems as a custom,
untracked (by git) mapping.