Nvim: turned visual mode into a two headed snake:
When in visual mode, not only can you move the "head" of the selection,
the normal way of changing what is selected, but you can also move the
"tail", or the other end of the what is connected, by pressing control
followed by a movement key. So far this only works for certain
pre-mapped movement keys, but it should be easy to add more. This works
regardless of whether you moved the head forward from, or backwards from
the tail, and whether you moved it above or below the tail.