Trying out leaving mouse partially enabled in Vim:
Having it enabled some of the time allows me to resize Vim windows with
the mouse, which is more helpful now that I have a massive monitor and
want to fully use it via multiple Vim windows. Having it disabled for
some modes enables me to access the right click menu of the terminal
application I am currently using.