Updated Vim configuration stuff involving testing php code:
It is now more configurable for different testing tool. This happened
as a result of changing it to work with a different testing tool, and
the end result seems more general/flexible to me.

Also changed the default test method generated in a new php test class
to have a return value by default.