Added new aswesom vim mapping:
<leader><leader>t creates a line below the current one that contains a
@psalm-trace on the variable that was under the cursor.

Similar to the way you can dumpt he variable under the cursor. Copied
the code and improved upon it such that if the cursor is at the very and
of the line and almost over a variable but not quite because it is over
a semicolon, it will still grab the variable just to the left of that.
Also improved that code for the dump thing.