Added vupdate alias:
This is a command that downloads the latest stable appimage of Neovim
from their Github downloads page to your ~/.local/bin/ directory and
then prints the SHA256 checcksum so you can compare it if you want. This
seems to work well without changing the execute permission if you
already had it setup correclty as it leaves it that way.