Nvim config: Better way to get fugitive to follow changes in current working directory:
The old way made no sense. Fugitive is very buffer based, it literally
associates each buffer with a particular git directory. The old way just
messed up which directory it associated the current buffer with, it
didn't just tell fugitive which git directory to use as that is not
possible.

This way a new buffer is opened whenever the working directory changes.