Nvim config improvements:
Redirecting command output to a ~/.vimout file. This helps if I have the
mouse enabled and can't select the output that shows up there.

Sourcing a maps and commands lua file to move more of my config into Lua
in an organized way.

Added mapping to edit the .vimout file.