Improved Neovim config:
Added ToggleMouse command that toggles the mouse off and on. This can be
helpful when I want to select some text with the mouse using the
terminal instead of Neovim, which can help for things not in a buffer
like command output. I also have a 4th mouse button set to a macro that
runs this command so that I can run it from my mouse hand while using
the mouse.

Added a mapping for <leader>i that runs ":lua " leaving the cursor there
where I can finish typing in some lua code to be run by my Neovim. I see
this as a Lua alternative to commands, since commands are actually just
a way to run vimscript and process it in your current Neovim instance.
So this is a way to run Lua code and process it in your running Neovim
instance. Lua > Vimscript.