nvim:
Removed DirChanged special handling. It wasn't working well in
conjunction with nvim-tree.lua. It's time I accepted that fugitive is
buffer based. Made change to nvim-tree.lua to make it work more to my
liking. It was only working the other way before as the start of
something I was developing to get it to play nice with my DirChanged
special handling.