nvim bug fixes:
Changed g:dreamy_developer to be either v:true or v:false in Vimscript,
and either true or false in Lua. In Vimscript if you set something to
v:true, it will be true in Lua, and v:false will be false. Also setting
a variable to true in lua will make in v:true in Vimscript. Setting a
variable to false in lua will make it v:false in vimscript.

Also renamed customBefore.lua to .beforeinit.lua. sourcing it instead of
doing it the lua way. This way if you :source init.vim, it will
reprocess .beforeinit.lua. This is important because it is mostly used
to set dreamy_developer to true. Otherwise, resourcing init.vim would
reset it to the default value. Also this way, the name can start with a
. and it will still work.