nvim: improved maps:
Moved the <leader>t mapping into lua. Made it work differently in VS
Code, showing the Explorer panel. Changed several mappings that executed
commands to use "<Cmd>" instead of ':'. This allows them to execute a
little faster. It also prevents Vim from switching modes unnecessarily.
It also prevents the command from appearing in the command history,
which is usually a good thing. This can allow <leader>. to continue to
repeat the last command without changing it to something you would not
want to repeat in that way. If you already have a mapping for that
command, you can probably just use the mapping instead. Some mappings
make sense to allow repeating with <leader>. and some do not. There is
also the option of using the repeat plugin to allow using . to repeat it
instead of repeating the command, but I haven't tried using these both
together yet (by both I mean both using "<Cmd>" instead of ":" and the
repeat plugin).