Fixed a strange Neovim issue:
For some reason in visual mode, when I press "<leader>l", which maps to
"$", instead of going to the end of the line, it goes past the end of
the line. This is very annoying as it visually selects the newline
character, and I always have to press "h" to go back one character. This
is not how this behaves in any of the other modes. So I changed the
mapping in that mode to adjust for it by going back one character using
"h"