Beginning to remove Neovim config from this repo.
My Neovim needs are too complex to store together in a single repo with my other dotfiles. I am going to separate it into 2 different Neovim config repos, one for use with vscode-neovim, and the other just for regular neovim. They will not be intended to be installed on the sam system, even though some may want to use Neovim and vscode-neovim on one system. That will not be supported by these repos.